Improve ingest exception handling and add test coverage.
I'm only testing a very simple case at the moment, when the URL to the source data is not valid. GDAL doesn't currently throw an exception if this happens, the reference to the dataset is nil. In the future, this could probably be further improved, but it's definitely good enough for now.