Added GET source handler and related cleanup.
This commit adds the notion of an inventory to the system, a table used to keep track of source data stored in various layer tables.

As part of this work, the "layers" table and namespace has been renamed to "registry" to make it clear that the table doesn't store raster data of any sort. Functions for storing raster data as chips have also been moved to a different namespace. A namespace with inventory functions has also been introduced.

Added an index on the tile column, that was specifically requested to support a use case. This requirement influences the design on the schema, all layer source's are store together. Removing a layer (and the source metadata) is non-trivial because of this...

Added more tests and helpers.

Cleaned up routes; they now reflect underlying implementation details less.

Added an exception handling middleware.

The core ingest code has also been refactored to improve readability.