Use a better grid scale factor value.
The grid scale factor has been inverted. With this change, the property is defined in terms of meters per grid unit, thus avoiding problems related to representing much smaller numbers as floating point values in configuration files.

I reverted a change to a test that I previously messed with while working on this problem. It was not obvious from the test output that the hash-value of a chip had changed because the point snap function was finding a different tile.