Remove JSON decoding of non-present chip summary.
Inventory queries may return a substantial number of sources. Instead of returning the summary of chips for each source, a reduced set of properties is returned.

The function that decodes JSON is not necessary and could create some confusion, because the non-present chip data is decoded to a null value.