Remove chip-specs synonym.
The resource created to maintain backwards compatability with Merlin is being removed; Merlin will be updated to work with a more sensible name.

The information contained within the registry isn't what a chip-spec used to be (because the geometry is now defined using a grid data-structure) so using the previous name doesn't make sense.